requirements.txt
# Основные библиотеки
pytest==7.4.0
requests==2.31.0

# Дополнительные плагины pytest
pytest-html==4.0.0
pytest-xdist>=3.0.0
pytest-cov>=4.0.0

# Для работы с JSON и данными
python-dateutil>=2.8.0

allure-pytest==2.13.0
python-dotenv==1.0.0